refactor: 로또 번호 객체화.
