+++ src/main/java/lotto/Lotto.java
@@ -1,6 +1,5 @@
 package lotto;
 
-import java.util.HashSet;
 import java.util.List;
 import java.util.Objects;
 import java.util.Set;
@@ -8,16 +7,19 @@ import java.util.TreeSet;
 
 public class Lotto {
 	private final static int SIZE_OF_LOTTO = 6;
-	private final static int LOTTO_MIN_NUMBER = 1;
-	private final static int LOTTO_MAX_NUMBER = 45;
 
-	private final Set<Integer> numbers;
+	private final Set<LottoNumber> numbers;
 
 	public Lotto(final List<Integer> numbers) {
-		this(new TreeSet<>(numbers));
+		Set<LottoNumber> numberSet = new TreeSet<>();
+		for (final Integer number : numbers) {
+			numberSet.add(new LottoNumber(number));
+		}
+		this.numbers = numberSet;
+		validate();
 	}
 
-	public Lotto(final Set<Integer> numbers) {
+	public Lotto(final Set<LottoNumber> numbers) {
 		this.numbers = numbers;
 		validate();
 	}
@@ -25,11 +27,6 @@ public class Lotto {
 	private void validate() {
 		if (numbers.size() != SIZE_OF_LOTTO) {
 			throw new IllegalArgumentException("로또의 번호는 6개 이다.");
-		}
-		for (final Integer number : numbers) {
-			if(number < LOTTO_MIN_NUMBER || number > LOTTO_MAX_NUMBER){
-				throw new IllegalArgumentException("로또의 숫자 범위는 1~45 이다.");
-			}
 		}
 	}
